Correction bug suppression exercice courant

+++ conf_delete.php
@@ -52,40 +52,43 @@
 		$id = $sPOST['id'];
         $annee = $sPOST['annee'];
 
-        // Suppression Exercice
-        $sql = "DELETE FROM exercice WHERE id = ?";
-        $req = $pdo->prepare($sql);
-        $req->execute(array($id));   
-        		
-        // On test si les information courante de l'utilisateur sont toujours valides sinon on les effaces
-        $sql2 = "SELECT * FROM user WHERE id = ?";
-        $req = $pdo->prepare($sql2);
-        $req->execute(array($user_id)); 
-        $data2 = $req->fetch(PDO::FETCH_ASSOC);
-        
-        if ($data2['exerciceid_encours'] == $id) { 
-            $sql3 = "UPDATE user SET exerciceid_encours = ?, mois_encours = ?  WHERE id = ?";
-            $req = $pdo->prepare($sql3);
-            $req->execute(array(NULL,NULL,$user_id));
-            // On supprimer la session ABODEP
-            $_SESSION['abodep'] = array();          
-        }        
-        
-        // On supprime la session
-        if ($annee == $exercice_annee) {
-            $_SESSION['exercice'] = array();
-        }
-
-        // Suppression depenses et recettes
-        $sql3 = "DELETE FROM recette WHERE exercice_id = ?";
-        $req = $pdo->prepare($sql3);
-        $req->execute(array($id));
-
-        $sql4 = "DELETE FROM depense WHERE exercice_id = ?";
-        $req = $pdo->prepare($sql4);
-        $req->execute(array($id));
-
-        // 	Suppression des paiements (par trigger)	
+		if ($exercice_annee != $annee) {
+			
+		    // Suppression Exercice
+		    $sql = "DELETE FROM exercice WHERE id = ?";
+		    $req = $pdo->prepare($sql);
+		    $req->execute(array($id));   
+		    		
+		    // On test si les information courante de l'utilisateur sont toujours valides sinon on les effaces
+		    $sql2 = "SELECT * FROM user WHERE id = ?";
+		    $req = $pdo->prepare($sql2);
+		    $req->execute(array($user_id)); 
+		    $data2 = $req->fetch(PDO::FETCH_ASSOC);
+			   
+	        if ($data2['exerciceid_encours'] == $id) { 
+	            $sql3 = "UPDATE user SET exerciceid_encours = ?, mois_encours = ?  WHERE id = ?";
+	            $req = $pdo->prepare($sql3);
+	            $req->execute(array(NULL,NULL,$user_id));
+	            // On supprimer la session ABODEP
+	            $_SESSION['abodep'] = array();          
+	        }        
+	        
+	        // On supprime la session
+	        if ($annee == $exercice_annee) {
+	            $_SESSION['exercice'] = array();
+	        }
+	
+	        // Suppression depenses et recettes
+	        $sql3 = "DELETE FROM recette WHERE exercice_id = ?";
+	        $req = $pdo->prepare($sql3);
+	        $req->execute(array($id));
+	
+	        $sql4 = "DELETE FROM depense WHERE exercice_id = ?";
+	        $req = $pdo->prepare($sql4);
+	        $req->execute(array($id));
+	
+	        // 	Suppression des paiements (par trigger)	
+		}
 	} 
 
 // On retourne d'ou on vient
